Fixed tests
Since upgrading to ruby 2.6.x there have been some tests failures. This commit
fixes this situation by upgrading some gems.

Moreover, since 1e5d2ac937c8fdfe6eac0250711c144f4fd91b56, the UrlValidator class
was not being called by the webhook model. I had to re-arrange things a bit to
make this part of the code work again.
